trigger_job/requirements.txt
google-cloud-run==0.10.7